app/src/main/java/gentle/hilt/learnunittests2/ui/ShoppingViewModel.kt
package gentle.hilt.learnunittests2.ui

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import dagger.hilt.android.lifecycle.HiltViewModel
import gentle.hilt.learnunittests2.db.ShoppingItem
import gentle.hilt.learnunittests2.others.Constants
import gentle.hilt.learnunittests2.others.Event
import gentle.hilt.learnunittests2.others.Resource
import gentle.hilt.learnunittests2.remote.reponses.ImageResponse
import gentle.hilt.learnunittests2.repository.ShoppingRepository
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class ShoppingViewModel @Inject constructor(
    private val repository: ShoppingRepository
) : ViewModel() {

    val liveDataShoppingItems = repository.observeAllShoppingItems()
    val liveDataTotalPrice = repository.observeTotalPrice()

    private val _images = MutableLiveData<Event<Resource<ImageResponse>>>()
    val images: LiveData<Event<Resource<ImageResponse>>> = _images

    private val _currentImageUrl = MutableLiveData<String>()
    val currentImageUrl: LiveData<String> = _currentImageUrl

    private val _insertShoppingItemStatus = MutableLiveData<Event<Resource<ShoppingItem>>>()
    val insertShoppingItemStatus: LiveData<Event<Resource<ShoppingItem>>> =
        _insertShoppingItemStatus


    fun setCurrentImageUrl(url: String) {
        _currentImageUrl.postValue(url)
    }

    fun deleteShoppingItem(shoppingItem: ShoppingItem) = viewModelScope.launch {
        repository.deleteShoppingItem(shoppingItem)
    }

    fun insertShoppingItemIntoDb(shoppingItem: ShoppingItem) = viewModelScope.launch {
        repository.insertShoppingItem(shoppingItem)
    }

    fun insertShoppingItem(name: String, amount: String, price: String) {
        if (name.isEmpty() || amount.isEmpty() || price.isEmpty()) {
            _insertShoppingItemStatus.postValue(Event(Resource.Error("The filed must not be empty")))
            return
        }
        if (name.length > Constants.MAX_NAME_LENGTH) {
            _insertShoppingItemStatus.postValue(Event(Resource.Error("Name of item " +
                    "must not exceed ${Constants.MAX_NAME_LENGTH} characters")))
            return
        }
        if(price.length > Constants.MAX_PRICE_LENGTH){
            _insertShoppingItemStatus.postValue(Event(Resource.Error("Price of item" +
            "must not exceed ${Constants.MAX_PRICE_LENGTH} characters")))
            return
        }
        val amount = try{
            amount.toInt()
        }catch (e:Exception){
            _insertShoppingItemStatus.postValue(Event(Resource.Error("Please enter a valid amount")))
            return
        }
        val shoppingItem = ShoppingItem(name, amount, price.toFloat(), currentImageUrl.value ?: "")
        insertShoppingItemIntoDb(shoppingItem)
        setCurrentImageUrl("")
        _insertShoppingItemStatus.postValue(Event(Resource.Success(shoppingItem)))
    }

    fun searchForImage(imageQuery: String) {
        if(imageQuery.isEmpty()){
            return
        }
        _images.value = Event(Resource.Loading(null))
        viewModelScope.launch {
            val response = repository.searchForImage(imageQuery)
            _images.value = Event(response)
        }

    }


}